+++ NEWS.md
@@ -1,4 +1,4 @@
-# svrep (dev)
+# svrep 0.6.4
 
 -   Fix issue #32: `stack_replicate_designs()` would only accept designs with types known to the 'survey' package. Fixed to allow other design types such as the random-groups jackknife.